double num survey point

+++ src/com/topodroid/DistoX/DrawingUtil.java
@@ -37,17 +37,17 @@ public class DrawingUtil
 
   // float toSceneX( float x, float y ) { return x; } 
   // float toSceneY( float x, float y ) { return y; } 
-  // public static float toSceneX( double x, double y ) { return (float)(CENTER_X + x * SCALE_FIX); }
-  // public static float toSceneY( double x, double y ) { return (float)(CENTER_Y + y * SCALE_FIX); }
-  public static float toSceneX( float x, float y ) { return (CENTER_X + x * SCALE_FIX); }
-  public static float toSceneY( float x, float y ) { return (CENTER_Y + y * SCALE_FIX); }
+  public static float toSceneX( double x, double y ) { return (float)(CENTER_X + x * SCALE_FIX); }
+  public static float toSceneY( double x, double y ) { return (float)(CENTER_Y + y * SCALE_FIX); }
+  // public static float toSceneX( float x, float y ) { return (CENTER_X + x * SCALE_FIX); }
+  // public static float toSceneY( float x, float y ) { return (CENTER_Y + y * SCALE_FIX); }
 
   // float sceneToWorldX( float x, float y ) { return x; } 
   // float sceneToWorldY( float x, float y ) { return y; }
-  // public static float sceneToWorldX( double x, double y ) { return (float)((x - CENTER_X)/SCALE_FIX); }
-  // public static float sceneToWorldY( double x, double y ) { return (float)((y - CENTER_Y)/SCALE_FIX); }
-  public static float sceneToWorldX( float x, float y ) { return ((x - CENTER_X)/SCALE_FIX); }
-  public static float sceneToWorldY( float x, float y ) { return ((y - CENTER_Y)/SCALE_FIX); }
+  public static float sceneToWorldX( double x, double y ) { return (float)((x - CENTER_X)/SCALE_FIX); }
+  public static float sceneToWorldY( double x, double y ) { return (float)((y - CENTER_Y)/SCALE_FIX); }
+  // public static float sceneToWorldX( float x, float y ) { return ((x - CENTER_X)/SCALE_FIX); }
+  // public static float sceneToWorldY( float x, float y ) { return ((y - CENTER_Y)/SCALE_FIX); }
     
   // int toBoundX( float x, float y ) { return Math.round(x); } 
   // int toBoundY( float x, float y ) { return Math.round(y); }
